Properly use build argument for app version in Dockerfile
This commit removes default value for Dockerfile arg `APP_VERSION` so
that builders know they need to pass it the app version value.
Additionally, it modifies the build task to reflect this and also to
automatically tag with "latest".
Finally, fix the repo URL which was pointing to the old address and bump
the Dockerfile version.